- refactor code base - setup flake8 for linting
