+++ app/templates/_post.html
@@ -7,7 +7,7 @@
         </td>
         <td>
             {% set user_link %}
-            <a href="{{ url_for('user', username=post.author.username) }}">
+            <a href="{{ url_for('main.user', username=post.author.username) }}">
                 {{ post.author.username }}
             </a>
             {% endset %}
