+++ app/templates/auth/login.html
@@ -9,9 +9,9 @@
     </div>
 </div>
 <br>
-<p>{{ _('New User?') }} <a href="{{ url_for('register') }}">{{ _('Click to Register!') }}</a></p>
+<p>{{ _('New User?') }} <a href="{{ url_for('auth.register') }}">{{ _('Click to Register!') }}</a></p>
 <p>
     {{ _('Forgot Your Password?') }}
-    <a href="{{ url_for('reset_password_request') }}">{{ _('Click to Reset It') }}</a>
+    <a href="{{ url_for('auth.reset_password_request') }}">{{ _('Click to Reset It') }}</a>
 </p>
 {% endblock %}
