+++ app/templates/base.html
@@ -18,30 +18,30 @@
         <span class="icon-bar"></span>
         <span class="icon-bar"></span>
       </button>
-      <a class="navbar-brand" href="{{ url_for('index') }}">Microblog</a>
+      <a class="navbar-brand" href="{{ url_for('main.index') }}">Microblog</a>
     </div>
     <div class="collapse navbar-collapse" id="bs-example-navbar-collapse-1">
       <ul class="nav navbar-nav">
         <li>
-          <a href="{{ url_for('index') }}">{{ _("Home") }}</a>
+          <a href="{{ url_for('main.index') }}">{{ _("Home") }}</a>
         </li>
         <li>
-          <a href="{{ url_for('explore') }}">{{ _("Explore") }}</a>
+          <a href="{{ url_for('main.explore') }}">{{ _("Explore") }}</a>
         </li>
       </ul>
       <ul class="nav navbar-nav navbar-right">
         {% if current_user.is_anonymous %}
         <li>
-          <a href="{{ url_for('login') }}">{{ _("Login") }}</a>
+          <a href="{{ url_for('auth.login') }}">{{ _("Login") }}</a>
         </li>
         {% else %}
         <li>
-          <a href="{{ url_for('user', username=current_user.username) }}">{{
+          <a href="{{ url_for('main.user', username=current_user.username) }}">{{
             _("Profile")
           }}</a>
         </li>
         <li>
-          <a href="{{ url_for('logout') }}">{{ _("Logout") }}</a>
+          <a href="{{ url_for('auth.logout') }}">{{ _("Logout") }}</a>
         </li>
         {% endif %}
       </ul>
