+++ app/templates/errors/404.html
@@ -2,5 +2,5 @@
 
 {% block app_content %}
 <h1>{{ _('Not Found') }}</h1>
-<p><a href="{{ url_for('index') }}">{{ _('Back') }}</a></p>
+<p><a href="{{ url_for('main.index') }}">{{ _('Back') }}</a></p>
 {% endblock %}
