+++ app/templates/user.html
@@ -15,17 +15,17 @@
       </p>
       {% if user == current_user %}
       <p>
-        <a href="{{ url_for('edit_profile') }}">{{ _("Edit your profile") }}</a>
+        <a href="{{ url_for('main.edit_profile') }}">{{ _("Edit your profile") }}</a>
       </p>
       {% elif not current_user.is_following(user) %}
       <p>
-        <a href="{{ url_for('follow', username=user.username) }}">{{
+        <a href="{{ url_for('main.follow', username=user.username) }}">{{
           _("Follow")
         }}</a>
       </p>
       {% else %}
       <p>
-        <a href="{{ url_for('unfollow', username=user.username) }}">{{
+        <a href="{{ url_for('main.unfollow', username=user.username) }}">{{
           _("Unfollow")
         }}</a>
       </p>
