+++ setup.cfg
@@ -1,0 +1,11 @@
+[flake8]
+ignore = E402, F401
+exclude =
+    .git,
+    __pycache__,
+    docs/source/conf.py,
+    old,
+    build,
+    dist
+max-line-length = 90
+max-complexity = 10
